requirements/base.txt
Django==1.5.4
django-tenant-schemas==1.4.1-2-g6407150
python-social-auth==0.1.17
